Reactive Workshop Step 1
Update the heroes service by creating input streams and behavior subjects with initial values that return search results, creating the search results stream, and deleting the search heroes method. Update the heroes page removing the search function and subscribing to the search results in the constructor. Update the hero search component by removing the search term and search results fields and properties, the timer, the searchHeroes function, and all places in the template that reference them, by subscribing to the heroes service search result stream in the constructor and storing the results, and by unsubscribing in ngOnDestroy.