Reactive Workshop Step 3
Update heroes service to debounce the search results stream, to reset the page number to zero when the search term or page size changes, to refresh the search results when a hero is added or deleted, and to share all streams and replay the last value.
Update the heroes page component class to stop subscribing to the search results stream in the constructor, to define disable previous and disable next page button streams, and to share all streams and replay the last value.
Update the heroes page template to use the heroes service search results stream and the heroes page disable previous and disable next page button streams.
Update the heroes detail page component with the appropriate types in the subscription to the get hero stream.